- Partially implemented call group logic + UI
Atm call groups are created but the checked calls are not hidden from the main table. Adding and removing calls from existing call groups has not been implemented.

TODO: Add filtering function on client side which hides rows that are in a call group.
TODO: Populate the dropdown of "Add calls to existing group" and write logic for it
TODO: Add click listener to remove row from call group button; Use Shiny.setInputValue to notify the server which row from which call group is to be removed.
TODO: Place accordion inside a scrolling panel?