Update dataLimits even if table is null
This fixes a bug where if the datatable has not been loaded, then the date limits where not getting updated properly.

This led to filtering function not hiding already matched calls properly, except if the user interacted first with the datetime slider, which managed to
update the datelimits since the table was not null anymore.