Only invalidate the cells of the timediff column on timediff update
This commit drastically improves the performance of the row selection logic to calculate the time differene.

Previously all the rows in the table were invalidated when the values of the timediff column were updated. Now, only the cells of the timediff column are invalidated
which drastically reduces the workload of datatables as it doesn't have to invalidate the cache for all columns and all rows.

Moreover, we restrict the deselect event function to run only when the user clicks on the last selected row.
Previously, it ran whenever any row was deselected which meant that the timediff column values were set to "-" BEFORE they were updated again to the correct
timediff values. This is because the deselect event fires BEFORE the select event. Hence, we were looping through all the rows TWICE => not our best work :(